// src/data/battleConstants.gen.ts
export interface BattleConstantMap {
  [constantName: string]: number;
}
// ...
export const B_OUTCOME: BattleConstantMap = {
  B_OUTCOME_WON: 1,
  B_OUTCOME_LOST: 2,
  B_OUTCOME_DREW: 3,
  B_OUTCOME_RAN: 4,
  B_OUTCOME_PLAYER_TELEPORTED: 5,
  B_OUTCOME_MON_FLED: 6,
  B_OUTCOME_CAUGHT: 7,
  B_OUTCOME_NO_SAFARI_BALLS: 8,
  B_OUTCOME_FORFEITED: 9,
  B_OUTCOME_MON_TELEPORTED: 10,
  B_OUTCOME_LINK_BATTLE_RAN: 128,
};